requirements.txt
flask==2.3.2
flask_cors==3.0.10
yt_dlp==2025.5.22
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
loguru==0.7.2